requirements.txt
boto
MySQL-Python